+++ catalog/app/model/graphql/schema.generated.ts
@@ -239,6 +239,10 @@ export default {
             kind: 'OBJECT',
             name: 'SnsInvalid',
           },
+          {
+            kind: 'OBJECT',
+            name: 'SubscriptionInvalid',
+          },
         ],
       },
       {
@@ -3596,6 +3600,18 @@ export default {
             ],
           },
           {
+            name: 'subscription',
+            type: {
+              kind: 'NON_NULL',
+              ofType: {
+                kind: 'OBJECT',
+                name: 'SubscriptionState',
+                ofType: null,
+              },
+            },
+            args: [],
+          },
+          {
             name: 'policies',
             type: {
               kind: 'NON_NULL',
@@ -4500,6 +4516,53 @@ export default {
             name: 'Unavailable',
           },
         ],
+      },
+      {
+        kind: 'OBJECT',
+        name: 'SubscriptionInvalid',
+        fields: [
+          {
+            name: '_',
+            type: {
+              kind: 'SCALAR',
+              name: 'Boolean',
+              ofType: null,
+            },
+            args: [],
+          },
+        ],
+        interfaces: [],
+      },
+      {
+        kind: 'OBJECT',
+        name: 'SubscriptionState',
+        fields: [
+          {
+            name: 'active',
+            type: {
+              kind: 'NON_NULL',
+              ofType: {
+                kind: 'SCALAR',
+                name: 'Boolean',
+                ofType: null,
+              },
+            },
+            args: [],
+          },
+          {
+            name: 'timestamp',
+            type: {
+              kind: 'NON_NULL',
+              ofType: {
+                kind: 'SCALAR',
+                name: 'Datetime',
+                ofType: null,
+              },
+            },
+            args: [],
+          },
+        ],
+        interfaces: [],
       },
       {
         kind: 'OBJECT',